Move project types to shared feature-projects
